chore(srcts): Separate _utils and web component helpers
