+++ srcts/src/components/_utils.ts
@@ -2,7 +2,6 @@ import type { HtmlDep } from "rstudio-shiny/srcts/types/src/shiny/render";
 
 import type { InputBinding as InputBindingType } from "rstudio-shiny/srcts/types/src/bindings/input";
 
-import { LitElement } from "lit";
 import type { BindScope } from "rstudio-shiny/srcts/types/src/shiny/bind";
 
 // Exclude undefined from T
@@ -75,127 +74,6 @@ function getAllFocusableChildren(el: HTMLElement): HTMLElement[] {
   return Array.from(focusable) as HTMLElement[];
 }
 
-class LightElement extends LitElement {
-  elementChildren: ChildNode[] = [];
-  slotContents: any;
-
-  static isShinyInput = false;
-
-  connectedCallback(): void {
-    this.elementChildren = Array.from(this.childNodes);
-    this.maybeCarryFill();
-    super.connectedCallback();
-  }
-
-  get slotElements(): ChildNode[] {
-    return this.elementChildren;
-  }
-
-  createRenderRoot(): this {
-    return this;
-  }
-
-  maybeCarryFill(): void {
-    if (this.isFillCarrier) {
-      this.classList.add("html-fill-container");
-      this.classList.add("html-fill-item");
-    } else {
-      this.classList.remove("html-fill-container");
-      this.classList.remove("html-fill-item");
-    }
-  }
-
-  get isFillCarrier(): boolean {
-    if (!this.parentElement) {
-      return false;
-    }
-
-    const inContainer = this.parentElement.classList.contains(
-      "html-fill-container"
-    );
-    const hasFillItem = Array.from(this.children).some((x: Element) =>
-      x.classList.contains("html-fill-item")
-    );
-
-    return inContainer && hasFillItem;
-  }
-}
-
-export interface CustomElementInputValue<T> extends HTMLElement {
-  id: string;
-  value: T;
-  onChangeCallback: (x: boolean) => void;
-  receiveMessage: (el: HTMLElement, data: { [key: string]: any }) => void;
-}
-
-export interface CustomElementInputGetValue<T> extends HTMLElement {
-  id: string;
-  getValue: () => T;
-  onChangeCallback: (x: boolean) => void;
-  receiveMessage: (el: HTMLElement, data: { [key: string]: any }) => void;
-}
-
-export type CustomElementInput<T> =
-  | CustomElementInputGetValue<T>
-  | CustomElementInputValue<T>;
-
-/**
- * Given a tag name for a custom element that is a CustomElementInput<T>, this
- * will hook up the proper input binding and register it with Shiny.
- * @param tagName Name of the tag that corresponds to the input binding
- * @returns Nothing
- */
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-function makeInputBinding<T>(
-  tagName: string,
-  { type = null }: { type?: string | null } = {}
-) {
-  if (!window.Shiny) {
-    return;
-  }
-
-  class NewCustomBinding extends Shiny["InputBinding"] {
-    constructor() {
-      super();
-    }
-
-    find(scope: HTMLElement): JQuery<CustomElementInput<T>> {
-      return $(scope).find(tagName) as JQuery<CustomElementInput<T>>;
-    }
-
-    getValue(el: CustomElementInputGetValue<T> | CustomElementInputValue<T>) {
-      if ("getValue" in el) {
-        return el.getValue();
-      } else {
-        return el.value;
-      }
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    getType(el: CustomElementInput<T>): string | null {
-      return type;
-    }
-
-    subscribe(el: CustomElementInput<T>, callback: (x: boolean) => void): void {
-      el.onChangeCallback = callback;
-    }
-
-    unsubscribe(el: CustomElementInput<T>): void {
-      // eslint-disable-next-line @typescript-eslint/no-empty-function, @typescript-eslint/no-unused-vars
-      el.onChangeCallback = (x: boolean) => {};
-    }
-
-    receiveMessage(
-      el: CustomElementInput<T>,
-      data: { [key: string]: any }
-    ): void {
-      el.receiveMessage(el, data);
-    }
-  }
-
-  Shiny.inputBindings.register(new NewCustomBinding(), `${tagName}-Binding`);
-}
-
 function bindAll(scope: BindScope): void {
   if (Shiny && Shiny.bindAll) Shiny.bindAll(scope);
 }
@@ -210,8 +88,6 @@ export {
   hasDefinedProperty,
   doWindowResizeOnElementResize,
   getAllFocusableChildren,
-  LightElement,
-  makeInputBinding,
   bindAll,
   unbindAll,
 };
